+++ src/main/java/ru/vova777/coordinatePlane/SecondCoordinateQuarter.java
@@ -4,32 +4,32 @@ public class SecondCoordinateQuarter extends CoordinateQuarter {
 
 
     @Override
-    public int getFinishSectionX(int x0, int length, int azimuth) {
-        int deltaX = (int) (Math.cos(Math.toRadians(180 - azimuth)) * length);
-        int x = x0 - deltaX;
+    public double getFinishSectionX(double x0, double length, double azimuth) {
+        double deltaX = (Math.cos(Math.toRadians(180 - azimuth)) * length);
+        double x = x0 - deltaX;
         return x;
     }
 
     @Override
-    public int getFinishSectionY(int y0, int length, int azimuth) {
-        int deltaY = (int) (Math.sin(Math.toRadians(180 - azimuth)) * length);
-        int y = y0 + deltaY;
+    public double getFinishSectionY(double y0, double length, double azimuth) {
+        double deltaY = (Math.sin(Math.toRadians(180 - azimuth)) * length);
+        double y = y0 + deltaY;
         return y;
     }
 
     @Override
-    public double getAzimuth(int delta_x, int lengthTrack) {
+    public double getAzimuth(double delta_x, double lengthTrack) {
         double angle = Math.toDegrees(Math.acos((double) delta_x / lengthTrack));
         return (180 - angle);
     }
 
     @Override
-    public int getDelta_x(int x0, int finish_x) {
+    public double getDelta_x(double x0, double finish_x) {
         return (x0 - finish_x);
     }
 
     @Override
-    public int getDelta_y(int y0, int finish_y) {
+    public double getDelta_y(double y0, double finish_y) {
         return (finish_y - y0);
     }
